LayoutTests/imported/w3c/web-platform-tests/dom/nodes/Document-contentType/contentType/contenttype_javascripturi-expected.txt


FAIL Javascript URI document.contentType === 'text/html' assert_equals: expected "text/html" but got ""
